chore: moved SortDataset to dedicated module
